SPT Frontend with table

+++ VsViewer/vs_calc/utils.py
@@ -11,17 +11,17 @@ def convert_to_midpoint(measures: np.ndarray, depths: np.ndarray):
         measure = measures[ix]
         if ix == 0:
             new_depths.append(0)
-            new_measures.append(measure)
+            new_measures.append(float(measure))
         else:
             if prev_depth is not None:
-                new_depths.append((depth + prev_depth) / 2)
-                new_measures.append(prev_measure)
-                new_depths.append((depth + prev_depth) / 2)
-                new_measures.append(measure)
+                new_depths.append(float((depth + prev_depth) / 2))
+                new_measures.append(float(prev_measure))
+                new_depths.append(float((depth + prev_depth) / 2))
+                new_measures.append(float(measure))
         if ix == len(depths) - 1:
             # Add extra depth for last value in array
-            new_depths.append(depth)
-            new_measures.append(measure)
+            new_depths.append(float(depth))
+            new_measures.append(float(measure))
         prev_depth = depth
         prev_measure = measure
     return new_measures, new_depths
